Refactor Ability Score File Layout
