+++ src/features/dialog-manager/dialogs/ability-dialog/abilities/constitution/actions/decrement-constitution.js
