+++ src/features/dialog-manager/dialogs/ability-dialog/abilities/intelligence/actions/decrement-intelligence.js
