+++ src/features/dialog-manager/dialogs/ability-dialog/abilities/intelligence/actions/increment-intelligence.js
