+++ src/features/dialog-manager/dialogs/ability-dialog/abilities/strength/actions/decrement-strength.js
