+++ src/features/dialog-manager/dialogs/ability-dialog/abilities/wisdom/actions/decrement-wisdom.js
